requirements.txt
matplotlib>=3.7.2
numpy>=1.25.1
setuptools>=65.5.0
scipy>=1.11.2
pytest>=7.4.2
black>=23.11.0
pillow>=10.3.0
pyparsing>=3.1.2
python-dateutil>=2.9.0.post0
scipy>=1.13.0
six>=1.16.0
cupy-cuda12x>=13.0.0
